+++ tests/unit/test_api.py
@@ -759,7 +759,9 @@ def test_remove(monkeypatch):
 
     def mock_post(path, body):
         assert path == "/snaps/placeholder"
-        assert body == {"action": "remove"}
+        assert body == {"action": "remove",
+                        "purge": False,
+                        "terminate": False}
 
         return mock_response
 
@@ -775,7 +777,9 @@ def test_remove_exception(monkeypatch):
 
     def mock_post(path, body):
         assert path == "/snaps/placeholder"
-        assert body == {"action": "remove"}
+        assert body == {"action": "remove", 
+                        "purge": False,
+                        "terminate": False}
 
         raise http.SnapdHttpException()
 
@@ -976,6 +980,100 @@ def test_list(monkeypatch):
     monkeypatch.setattr(http, "get", mock_get)
 
     result = api.list()
+
+    assert result == mock_response
+
+
+def test_list_all(monkeypatch):
+    """`api.list_all` returns a `types.SnapdResponse`."""
+    mock_response = types.SnapdResponse(
+        type="sync",
+        status_code=200,
+        status="OK",
+        result=[{"title": "placeholder1"}, {"title": "placeholder2"}],
+    )
+
+    def mock_get(path):
+        assert path == "/snaps?select=all"
+
+        return mock_response
+
+    monkeypatch.setattr(http, "get", mock_get)
+
+    result = api.list_all()
+
+    assert result == mock_response
+
+
+def test_snapshots(monkeypatch):
+    """`api.snapshots` returns a `types.SnapdResponse`."""
+    mock_response = types.SnapdResponse(
+        type="sync",
+        status_code=200,
+        status="OK",
+        result=[{"title": "placeholder1"}, {"title": "placeholder2"}],
+    )
+
+    def mock_get(path):
+        assert path == "/snapshots"
+
+        return mock_response
+
+    monkeypatch.setattr(http, "get", mock_get)
+
+    result = api.snapshots()
+
+    assert result == mock_response
+
+
+def test_save_snapshot(monkeypatch):
+    """`api.snapshots` returns a `types.SnapdResponse`."""
+    mock_response = types.SnapdResponse(
+        type="async",
+        status_code=202,
+        status="OK",
+        result=[{"title": "placeholder1"}, {"title": "placeholder2"}],
+    )
+
+    def mock_post(path, body):
+        assert path == "/snaps"
+        assert body == {
+            "action": "snapshot",
+            "snaps": ["snapd"],
+            "users": ["user1"],}
+        
+        return mock_response
+
+    monkeypatch.setattr(http, "post", mock_post)
+
+    result = api.save_snapshot(snaps=["snapd"], users=["user1"])
+
+    assert result == mock_response
+
+
+def test_forget_snapshot(monkeypatch):
+    """`api.forget_snapshot` returns a `types.SnapdResponse`."""
+    mock_response = types.SnapdResponse(
+        type="async",
+        status_code=202,
+        status="OK",
+        result=[{"title": "placeholder1"}, {"title": "placeholder2"}],
+    )
+
+    def mock_post(path, body):
+        assert path == "/snapshots"
+        assert body == {
+            "action": "forget",
+            "set": 1,
+            "snaps": ["snapd"],
+            "users": ["user1"],
+            }
+
+        return mock_response
+
+    monkeypatch.setattr(http, "post", mock_post)
+
+    result = api.forget_snapshot(1, snaps=["snapd"], users=["user1"])
 
     assert result == mock_response